+++ eth_defi/erc_4626/vault_protocol/yieldnest/__init__.py
@@ -1,0 +1,1 @@
+"""YieldNest protocol integration."""